+++ changelog.md
@@ -1,3 +1,4 @@
 - Fixed fluid map colors being incorrect.
 - Fixed crash on startup when Architectury API is not installed due to accessing protected TransparentBlock class.
 - Space station recipes now specify the nbt structure file. This allows you to set different space station structures for different planets.
+- Added config to disable Air Vortexes.